Move columsn to parent component
