+++ src/sections/evidence/EVA/EVAClientTable.js
@@ -1,158 +1,11 @@
 import React from 'react';
-import { Link as RouterLink } from 'react-router-dom';
 import { gql, useQuery } from '@apollo/client';
-import { Link } from 'ot-ui';
 import SectionItem from '../../../components/Section/SectionItem';
 import { betaClient } from '../../../client';
-import { clinvarStarMap, naLabel } from '../../../constants';
 import usePlatformApi from '../../../hooks/usePlatformApi';
 import Summary from './Summary';
-import { DataTable, TableDrawer } from '../../../components/Table';
-import ClinvarStar from '../../../components/ClinvarStar';
-import Tooltip from '../../../components/Tooltip';
+import { DataTable } from '../../../components/Table';
 import Description from './Description';
-import { epmcUrl } from '../../../utils/urls';
-
-const columns = [
-  {
-    id: 'disease.name',
-    label: 'Disease/phenotype',
-    renderCell: ({ disease }) => {
-      return (
-        <Link component={RouterLink} to={`/disease/${disease.id}`}>
-          {disease.name}
-        </Link>
-      );
-    },
-  },
-  {
-    id: 'diseaseFromSource',
-    label: 'Reported disease/phenotype',
-  },
-  {
-    id: 'variantRsId',
-    label: 'Variant',
-    renderCell: ({ variantRsId }) => {
-      return variantRsId ? (
-        <Link
-          external
-          to={`http://www.ensembl.org/Homo_sapiens/Variation/Explore?v=${variantRsId}`}
-        >
-          {variantRsId}
-        </Link>
-      ) : (
-        naLabel
-      );
-    },
-  },
-  {
-    id: 'studyId',
-    label: 'ClinVar ID',
-    renderCell: ({ studyId }) => {
-      return studyId ? (
-        <Link external to={`https://www.ncbi.nlm.nih.gov/clinvar/${studyId}`}>
-          {studyId}
-        </Link>
-      ) : (
-        naLabel
-      );
-    },
-  },
-  {
-    label: 'Functional consequence',
-    renderCell: ({ variantFunctionalConsequence }) => {
-      return (
-        <Link
-          external
-          to={`http://www.sequenceontology.org/browser/current_svn/term/${
-            variantFunctionalConsequence.id
-          }`}
-        >
-          {variantFunctionalConsequence.label}
-        </Link>
-      );
-    },
-  },
-  {
-    id: 'clinicalSignificances',
-    label: 'Clinical significance',
-    renderCell: ({ clinicalSignificances }) => {
-      return !clinicalSignificances ? (
-        naLabel
-      ) : clinicalSignificances.length === 1 ? (
-        clinicalSignificances[0]
-      ) : (
-        <ul
-          style={{
-            margin: 0,
-            padding: 0,
-            listStyle: 'none',
-          }}
-        >
-          {clinicalSignificances.map(clinicalSignificance => {
-            return <li key={clinicalSignificance}>{clinicalSignificance}</li>;
-          })}
-        </ul>
-      );
-    },
-  },
-  {
-    id: 'allelicRequirements',
-    label: 'Allelic requirement',
-    renderCell: ({ allelicRequirements }) => {
-      return !allelicRequirements ? (
-        naLabel
-      ) : allelicRequirements.length === 1 ? (
-        allelicRequirements[0]
-      ) : (
-        <ul
-          style={{
-            margin: 0,
-            padding: 0,
-            listStyle: 'none',
-          }}
-        >
-          {allelicRequirements.map(allelicRequirement => {
-            return <li key={allelicRequirement}>{allelicRequirement}</li>;
-          })}
-        </ul>
-      );
-    },
-  },
-  {
-    label: 'Confidence',
-    renderCell: ({ confidence }) => {
-      const numStars = clinvarStarMap[confidence];
-      const stars = [];
-      for (let i = 0; i < numStars; i++) {
-        stars.push(<ClinvarStar key={i} />);
-      }
-      return (
-        <Tooltip title={confidence}>
-          <span>{stars}</span>
-        </Tooltip>
-      );
-    },
-  },
-  {
-    label: 'Literature',
-    renderCell: ({ literature }) => {
-      const literatureList =
-        literature?.reduce((acc, id) => {
-          if (id !== 'NA') {
-            acc.push({
-              name: id,
-              url: epmcUrl(id),
-              group: 'literature',
-            });
-          }
-          return acc;
-        }, []) || [];
-
-      return <TableDrawer entries={literatureList} caption="Literature" />;
-    },
-  },
-];
 
 const EVA_QUERY = gql`
   query evaQuery($ensemblId: String!, $efoId: String!, $size: Int!) {
@@ -186,7 +39,7 @@ const EVA_QUERY = gql`
   }
 `;
 
-function EVAClientTable({ definition, id, label }) {
+function EVAClientTable({ definition, id, label, columns }) {
   const { ensgId: ensemblId, efoId } = id;
   const { data: summaryData } = usePlatformApi(Summary.fragments.evaSummary);
